some renamings
